fix: better control of opsdata encryption
Remove the `Enabled` attribute for baseline opsdata and rely instead on
the explicit `Enabled` attribute in the configuration data.

As there is now an explicit `Enabled` attribute for `baselinedata`
encryption with a default of `false`, there is no need for this attribute
to be included in the masterdata.

Removing the explicit value from the masterdata will also mean it is
controllable via deployment profiles. At present, the masterdata is
incorporated into the blueprint so can only be overridden in the
solution and not via a deployment profile.