fix(s3): do not validate replica sequence on delete
When replicating a data bucket into another, we only want
to validate the order of operations on their creation/update.
Performing thet same validation on the delete of a bucket
implies that a replica destination should still exist during
the deletion of the source - but that is not usually the req.