feat: input seeders
Implement seeders to provide input state via the refactored input source
processing.

Masterdata is now dynamically loaded directly from JSON files, and
provided to the masterdata stage. If is filtered by region if provided.

It also provides an example of a seeder that is specific to a particular
input source. The example needs to adjust the command line options but
only for the `mock` input source.

Note that this change has not affected the existing input processing. A
subsequent change will be introduced to perform the transition.