refactor: command line and masterdata access
Replace direct access to the command line options with use of accessor
functions. This will then permit the implementation of command line
options to be refactored to the new input source approach without
affecting this code.

Switch to using the input source process to determine the masterdata
available.

Update the AWS masterdata refresh process to update the regions.json
file in the inputseeder.