fix: handle naming changes for alerts
Some extra handling around naming to support replace templates with
alerts. Since the alerts rely on the ref to the queue they currently
fail to move during replacement. Instead we provide a fixed value for
the alert dimension to ensure we don't rename the alert and keep
monitoring in the same state during replacement