refactor: command line and masterdata access (#255)
* refactor: command line and masterdata access

Replace direct access to the command line options with use of accessor
functions. This will then permit the implementation of command line
options to be refactored to the new input source approach without
affecting this code.

Switch to using the input source process to determine the masterdata
available.

Update the AWS masterdata refresh process to update the regions.json
file in the inputseeder.

* refactor: CLO accessor names

All command line option accessor functions now start with `getCLO`.

Seeder functions have been renamed to avoid confusing terminology
(outputs as as part of input) referring instead to the `config` and
`state` pipelines.