refactor: composite template inclusion
Support composite templates to be provided via a filename or via the
content being directly included. Switch to the content based input
variable for composite accounts rather than the filename based
variable.

As part of moving to dynamic input loading, reliance on pre-assembled
content needs to be removed. This change starts the process by using
the `?interpret` builtin internally to avoid use of the `[#include]`
directive if the content appears to be freemarker markup.