feat(ec2): adds support for autoscale lifecycles
Adds support for autoscale lifecycle hooks focussing on the startup
process. This is similar to CFN init but handles the startup instances
as part of scaling activites. The instance will be started and if a
failure occurs or 10 minute timeout the scaling event will terminate the
instance and start again

Adds a default lifecycle hook for startup events with the ability to
override that if required.

Adds IAM policy to allow the instances to query the ASG they belong to
and send the event