feat: adds support for healthchecks
Implements simple and complex healthchecks as part of the healthcheck
component.

The simple check is performed by Route53 and runs distributed basic HTTP
checks against a single endpoint. The check has to be run from at least
3 regions that support healthchecks and uses our standard port
configuration to define the configuration of the check. A link to a
destination is required for the check to work.

The complex health check uses CloudWatch canary monitoring which uses
lambda to perform scripted synthetic transactions using browser
automation tools like selenium. The checks are deployed to a local
region and run at a schedule