feat(apigateway): opeanpi fragment and vpclink
Adds support for vpclinks between apigateway and nlb based on LB port
links. A link to a network lb port will add a priviateHTTPEndpoint
configuration and stage variable that is used during integration
processing to include the vpclink configuration

Adds support for defining openapi specification as part of a fragment
file. The base specification is still the spec provided via the image
source but this allows for it to be modified or defined within the
solution