fix: API Gateway Schema naming constraint
Ensure schema model names are alphanumeric.

Add support for modifying swagger specs to meet the additional
constraints applied by AWS. Much of this was previously done in the
bash executor using jq, but it seems more appropriate to apply these
constraints at the point of deployment rather than at the time of build.

Once this PR is merged, a subsequent PR will be put through to remove
the AWS specific processing in the bash executor.