feat(districts): Refactoring to support feat(districts)
Use accessor methods for a number of variables that were previously
global variables provided set setContext.

SetContext had a number of AWS specific lookups which only work
if the input state relates to an AWS deployment. Using accessors
defers the calculations until this situation is in place.

A future refactor would be to change the accessors to work off the
blueprint and in many cases, rework access once baseline access
is done via locations.