refactor: delivery stream encrpytion changes (#457)
Add explicit config to indicate when a Kinesis delivery stream is NOT
performing KMS encryption.

Kinesis does not gracefully handle the removal of encryption settings
on a delivery stream (even though the documented default value would
suggest it does). Instead, additional config is needed to indicate that
KMS encryption is explicitly NOT required.

This is an unusual situation as once set, encryption settings are
normally not altered. However, as cloudformation thinks everything is
ok (it hasn't forwarded the encryption options), to avoid confusion
and additional troubleshooting and frustration, the extra configuration
is explicitly set.