feat: kinesis dynamic partitioning
Add configuration support for enabling partitioning when the destination
is S3.

Dynamic partitioning must be enabled if adding delimiters between records
or when using dynamic partition operators in the prefix or error prefix.

Partition values will typically be added via a `metadata` attribute on
each record by a lambda processor.