feat(kinesis): Finer grained partition control
- add solution configuration to permit control on the use of dynamic
partitioning
- check conditions where partitioning should/shouldn't be enabled
- explicitly add line endings to firehose records for the api gateway,
removing the need to use delimiters and partitioning if just directing
logs to S3
- add config to the API gateway to permit control over the syntax used
to format log records (text/json) - this applies to cloudwatch and
firehose logs
- set defaults so existing  firhose deployments don't break, noting that
partition settings can only be set when a firehose is created.

A subsequent change will refine the informaiton included in the API
Gateway logs reflecting the functionality employed (waf, authorizer,
integration).