feat(apigateway): Execution Log Control (#476)
* feat(apigateway): Execution Log Control

Support subscriptions and retention control on the execution log.

The primary driver for this change is to permit the inclusion of
execution logs in the log consolidation process via a log subscription.

Because the api gateway does not permit control over the execution log
group name and includes the api id in the name it creates, special
logic is required to control log group retention (normally set when
creating the log group) and to create log subscriptions on the log group,
for instance for log consolidation.

The change creates a pseudo stack for the execution log group and
sets its retention according to the operational expiration setting.

In order that subscription processing works the same as if the execution
log group could be created, the deployment unit of the pseudo stack is
set to match what is used for the access log group. This also makes
for a straightforward transition in the future to explicit control of
the log group creation if/when AWS supports it.

* refactor(apigateway): ensure stack is run twice

Add a warning to ensure the stack is run twice if log subscriptions on
the execution log are required.