fix: s3 versioning without lifecycle management
Because of the location of the configuration controlling versioning of
the s3 bucket, if versioning is turned on, the bucket will automatically
be lifecycled.

To permit versioning without lifecycle management, support a separate
versioning attribute.

The Versioning attribute on Lifecycle will be deprecated to avoid
lifecycle management as an unwanted side effect of enabling versioning.