feat(s3): backup support (#516)
Add ability for an S3 bucket to be included in backups via
a link to a backup regime. This results in it receiving
the tag associated with the regime.

Also remove the limitation on the tags per bucket. There is a
limit of 10 tags per bucket object, but a bucket itself can have
up to 50 tags (https://docs.aws.amazon.com/general/latest/gr/aws_tagging.html).