feat(spa): force max-age for config
In order to ensure browsers don't excessivley cache the
spa config.json file, give it a max-age in s3 which will
flow through CloudFront to the browser.

Because the config.json typically contains the version
information displayed by the spa, it needs to be fetched
when new code is loaded to avoid the appearance that no
update has ocurre.