feat: policy chunking
When inline policies exceed the aggregate limit, convert
then to managed policies and optionally split them into
multiple statements if the managed policy size limit
would be exceeded.

The process some refactoring of the calling code so in
has just been done for lambda to bed in any issues. The bulk
of the work is done by general policy routines which can be applied
to other components as needed.