feat(lambda): versioned lambda retention policy
When a new version is created on each deploy, provide
control over whether the old versions are deleted or retained.

The default is to retain in line with the previous hard coded value.

The need to delete has become more relevant as versioned
lambdas are needed to preprovision lambda instances.