refactor(apigateway): authorization models
Support additional model values for the case of IP filtering in
combination with a lambda authorizer. Also rename the config attribute
to more correctly reflect its purpose in controlling authorization
rather than authentication.

When used with a lambda authorizer, the default value of "IP" incorrectly
provides an explict ALLOW rather than relying on it to come from the policy
provided by the authorizer. By providing explicit values to be used with the
authorizer, the configuration can be validated as appropriate.