refactor(apigateway): authorization models (#581)
* refactor(apigateway): authorization models

Support additional model values for the case of IP filtering in
combination with a lambda authorizer. Also rename the config attribute
to more correctly reflect its purpose in controlling authorization
rather than authentication.

When used with a lambda authorizer, the default value of "IP" incorrectly
provides an explict ALLOW rather than relying on it to come from the policy
provided by the authorizer. By providing explicit values to be used with the
authorizer, the configuration can be validated as appropriate.

* refactor(apigateway): authorisation model misconfiguration now fatal

Force the user to correct misconfigurations of the authorization model
if use of authorizers is detected.

* refactor(apigateway): prefix sig4 auth models

Add AWS specific values for authroisation models involving SIG4.

This will permit the removal of the non-preofixed versions in a future
breaking change.

* refactor(apigateway): move legacy auth models

Move the non-prefixed SIG4 authorization models to the AWS provider
as they are very specific to AWS. A future change can remove the
non-prefixed ones, though moving them to the AWS provider largely
isolates them anyway.