fix(apigateway): OPTIONS handling with authorisers
Add a statement to the resource policy to provide an explicit
ALLOW when the OPTIONS verb is requested and the response is
handled inline by the API Gateway.

If OPTIONS are handled inline by the API Gateway, they do not
contribute the explicit ALLOW that an authorizer would. In this
case, the resource policy must provide an explicit allow for the
API Gateway authorization model to succeed.