feat(lambda): lambda aliases
When deploying a fixed lambda version, provide the
option of a lambda alias to avoid the need to update
references when a new version of the lambda is created.

Provisioning is also applied to the alias if configured
in order that it moves to the latest version of the lambda.