refactor(iam): remove transitional policy support
When settings was switched to using the component raw name, code
was added to support a transition from the old paths to the new.

Unfortunately the transitional code did not include the `settings/`
prefix at the start of the path, meaning that the permissions did not
the previous format.

As a result the raw name based permissions have been the only effective ones
for a number of months.

This change removes the transitional code, mainly to reduce the
size of iam based passes, which otherwise have duplicate sections for the
standard iam permissions.