fix(image): include tag state from output
Currently the stack output generated by the executor uses the NAME
attribute while the image state code expects the TAG attribute.

In order to ensure the tag is included in the image state, and to
permit future rationalisation of the attribute used, support the use of
either attribute to provide the tag value.