refactor(datafeed): undeployed lambda functions
Include any lambda function processor definitions even if the lambda
function isn't deployed.

AWS requires that a lambda processor is configured if the prefix
definitions include lambda based partitioning.

To permit the datafeed to be deployed at the solution level before the
lambda has been deployed, the lambda processor configuration needs to be
included even if the lambda function isn't deployed.

This also adds the necessary iam permissions that will take effect as
soon as the lambda function is deployed.